chore(clientDb): renaming tables to singular
