+++ packages/client-db/index.ts
@@ -1,14 +1,14 @@
 export * from "./src/db";
 
-export * from "./src/emails/bulk-delete-emails";
-export * from "./src/emails/bulk-get-threads";
-export * from "./src/emails/bulk-put-emails";
-export * from "./src/emails/bulk-update-emails";
-export * from "./src/emails/get-email-threads-from-sender";
-export * from "./src/emails/update-email-sync-info";
-export * from "./src/emails/upsert-email-sync-info";
+export * from "./src/email/bulk-delete-emails";
+export * from "./src/email/bulk-get-threads";
+export * from "./src/email/bulk-put-emails";
+export * from "./src/email/bulk-update-emails";
+export * from "./src/email/get-email-threads-from-sender";
+export * from "./src/email/update-email-sync-info";
+export * from "./src/email/upsert-email-sync-info";
 
-export * from "./src/threads/bulk-put-threads";
+export * from "./src/thread/bulk-put-threads";
 
 export * from "./src/hooks/use-email-sync-info";
 export * from "./src/hooks/use-email-thread";
