+++ packages/client-db/src/email/get-email-sync-info.ts
